requirements.txt
pandas
streamlit
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
requests
openpyxl
gspread
oauth2client
